# server/src/modules/agents/agents.routes.ts
import type { FastifyInstance } from "fastify";
import type { ZodTypeProvider } from "fastify-type-provider-zod";
import { authGuard } from "../../shared/middleware/authGuard.js";
import {
  agentDetailSchema,
  agentPageSchema,
  agentSchema,
  apiKeyPageSchema,
  createAgentSchema,
  createCriterionSchema,
  createdKeySchema,
  createKeySchema,
  criterionParam,
  criterionSchema,
  idParam,
  keyParam,
  listAgentsQuery,
  listKeysQuery,
  okSchema,
  rubricSchema,
  runListPageSchema,
  runsQuery,
  updateAgentSchema,
  updateCriterionSchema,
  updateRubricSchema,
  versionLabelParam,
  versionSchema,
} from "./agents.schema.js";
import * as service from "./agents.service.js";
import { z } from "zod";

export function agentRoutes(app: FastifyInstance): void {
  const r = app.withTypeProvider<ZodTypeProvider>();
  r.addHook("preHandler", authGuard);

  r.get(
    "/",
    { schema: { querystring: listAgentsQuery, response: { 200: agentPageSchema } } },
    async (req) => {
      return service.list(req.user.sub, req.query);
    },
  );

  r.post(
    "/",
    { schema: { body: createAgentSchema, response: { 201: agentDetailSchema } } },
    async (req, reply) => {
      const agent = await service.create(req.user.sub, req.body);
      return reply.code(201).send(agent);
    },
  );

  r.get("/:id", { schema: { params: idParam, response: { 200: agentDetailSchema } } }, async (req) => {
    return service.getDetail(req.params.id, req.user.sub);
  });

  r.patch(
    "/:id",
    { schema: { params: idParam, body: updateAgentSchema, response: { 200: agentSchema } } },
    async (req) => {
      return service.rename(req.params.id, req.user.sub, req.body);
    },
  );

  r.delete("/:id", { schema: { params: idParam, response: { 200: okSchema } } }, async (req) => {
    await service.remove(req.params.id, req.user.sub);
    return { ok: true };
  });

  r.get(
    "/:id/rubric",
    { schema: { params: idParam, response: { 200: rubricSchema } } },
    async (req) => {
      return service.getRubric(req.params.id, req.user.sub);
    },
  );

  r.patch(
    "/:id/rubric",
    { schema: { params: idParam, body: updateRubricSchema, response: { 200: rubricSchema } } },
    async (req) => {
      return service.editRubric(req.params.id, req.user.sub, req.body);
    },
  );

  r.post(
    "/:id/rubric/criteria",
    { schema: { params: idParam, body: createCriterionSchema, response: { 201: criterionSchema } } },
    async (req, reply) => {
      const criterion = await service.addCriterion(req.params.id, req.user.sub, req.body);
      return reply.code(201).send(criterion);
    },
  );

  r.patch(
    "/:id/rubric/criteria/:cid",
    {
      schema: {
        params: criterionParam,
        body: updateCriterionSchema,
        response: { 200: criterionSchema },
      },
    },
    async (req) => {
      return service.editCriterion(req.params.cid, req.params.id, req.user.sub, req.body);
    },
  );

  r.delete(
    "/:id/rubric/criteria/:cid",
    { schema: { params: criterionParam, response: { 200: okSchema } } },
    async (req) => {
      await service.removeCriterion(req.params.cid, req.params.id, req.user.sub);
      return { ok: true };
    },
  );

  r.post(
    "/:id/keys",
    { schema: { params: idParam, body: createKeySchema, response: { 201: createdKeySchema } } },
    async (req, reply) => {
      const key = await service.issueKey(req.params.id, req.user.sub, req.body);
      return reply.code(201).send(key);
    },
  );

  r.get(
    "/:id/keys",
    { schema: { params: idParam, querystring: listKeysQuery, response: { 200: apiKeyPageSchema } } },
    async (req) => {
      return service.listKeys(req.params.id, req.user.sub, req.query);
    },
  );

  r.delete(
    "/:id/keys/:kid",
    { schema: { params: keyParam, response: { 200: okSchema } } },
    async (req) => {
      await service.revokeKey(req.params.kid, req.params.id, req.user.sub);
      return { ok: true };
    },
  );

  r.get(
    "/:id/runs",
    { schema: { params: idParam, querystring: runsQuery, response: { 200: runListPageSchema } } },
    async (req) => {
      return service.listRuns(req.params.id, req.user.sub, req.query);
    },
  );

  r.get(
    "/:id/versions",
    { schema: { params: idParam, response: { 200: z.array(versionSchema) } } },
    async (req) => {
      return service.listVersions(req.params.id, req.user.sub);
    },
  );

  r.get(
    "/:id/versions/:label",
    { schema: { params: versionLabelParam, response: { 200: versionSchema } } },
    async (req) => {
      return service.getVersion(req.params.id, req.user.sub, req.params.label);
    },
  );
}